requirements.txt
selenium==4.12.0
pytest==8.3.4
pytest-html==3.2.0
pytest-xdist==3.4.0
pyyaml==6.0